+++ LazyLoot/Ui/ConfigUi.cs
@@ -2,6 +2,7 @@ using Dalamud.Interface.Colors;
 using Dalamud.Interface.Internal.Notifications;
 using Dalamud.Interface.Windowing;
 using ImGuiNET;
+using LazyLoot.Services;
 using System;
 using System.Numerics;
 
@@ -13,7 +14,7 @@ namespace LazyLoot.Ui
         internal WindowSystem windowSystem = new();
         private string? toastPreview;
 
-        public ConfigUi() : base("Lazy Loot Config", ImGuiWindowFlags.AlwaysAutoResize)
+        public ConfigUi() : base("Lazy Loot Config" )
         {
             SizeConstraints = new WindowSizeConstraints()
             {
@@ -21,53 +22,73 @@ namespace LazyLoot.Ui
                 MaximumSize = new Vector2(99999, 99999),
             };
             windowSystem.AddWindow(this);
-            Service.Service.PluginInterface.UiBuilder.Draw += windowSystem.Draw;
+            Service.PluginInterface.UiBuilder.Draw += windowSystem.Draw;
         }
 
         public void Dispose()
         {
-            Service.Service.PluginInterface.UiBuilder.Draw -= windowSystem.Draw;
+            Service.PluginInterface.UiBuilder.Draw -= windowSystem.Draw;
             GC.SuppressFinalize(this);
         }
 
         public override void Draw()
         {
-            DrawFeatures();
-            DrawRollingDelay();
-            DrawChatAndToast();
-            DrawUserRestriction();
-            DrawFulf();
-            DrawSave();
+            if (ImGui.BeginTabBar("config"))
+            {
+                if (ImGui.BeginTabItem("Features"))
+                {
+                    ImGui.BeginChild("generalFeatures");
+
+                    DrawFeatures();
+                    DrawRollingDelay();
+                    DrawChatAndToast();
+                    DrawFulf();
+
+                    ImGui.EndChild();
+                    ImGui.EndTabItem();
+                }
+
+                if (ImGui.BeginTabItem("User Restriction"))
+                {
+                    ImGui.BeginChild("generalFeatures");
+
+                    DrawUserRestriction();
+
+                    ImGui.EndChild();
+                    ImGui.EndTabItem();
+                }
+
+                ImGui.EndTabBar();
+            }
+
         }
 
         public override void OnClose()
         {
             Plugin.LazyLoot.config.Save();
-            Service.Service.PluginInterface.UiBuilder.AddNotification("Configuration saved", "Lazy Loot", NotificationType.Success);
+            Service.PluginInterface.UiBuilder.AddNotification("Configuration saved", "Lazy Loot", NotificationType.Success);
             base.OnClose();
         }
 
         private static void DrawFeatures()
         {
-            ImGui.Text("Features");
-            ImGui.Separator();
-            ImGui.Text("/roll need");
+            ImGui.Text("/rolling need");
             ImGui.SameLine();
             ImGui.Text("Roll need for everything. If impossible roll greed or pass if greed is impossible.");
             ImGui.Separator();
-            ImGui.Text("/roll needonly");
+            ImGui.Text("/rolling needonly");
             ImGui.SameLine();
             ImGui.Text("Roll need for everything. If impossible, roll pass.");
             ImGui.Separator();
-            ImGui.Text("/roll greed");
+            ImGui.Text("/rolling greed");
             ImGui.SameLine();
             ImGui.Text("Roll greed for everything. If impossible, roll pass.");
             ImGui.Separator();
-            ImGui.Text("/roll pass");
+            ImGui.Text("/rolling pass");
             ImGui.SameLine();
             ImGui.Text("Pass on things you haven't rolled for yet.");
             ImGui.Separator();
-            ImGui.Text("/roll passall");
+            ImGui.Text("/rolling passall");
             ImGui.SameLine();
             ImGui.Text("Passes on all, even if you rolled on them previously.");
             ImGui.Separator();
@@ -76,38 +97,22 @@ namespace LazyLoot.Ui
         private static void DrawRollingDelay()
         {
             ImGui.Checkbox("Rolling delay between items:", ref Plugin.LazyLoot.config.EnableRollDelay);
-            ImGui.SameLine();
-            ImGui.SetNextItemWidth(100);
             if (Plugin.LazyLoot.config.EnableRollDelay)
             {
+                ImGui.SameLine();
+                ImGui.SetNextItemWidth(100);
                 ImGui.DragFloat("in seconds.", ref Plugin.LazyLoot.config.RollDelayInSeconds, 0.1F);
 
                 if (Plugin.LazyLoot.config.RollDelayInSeconds < 0.1f)
                 {
                     Plugin.LazyLoot.config.RollDelayInSeconds = 0.1f;
                 }
-
-                ImGui.Separator();
             }
-        }
-
-        private static void DrawSave()
-        {
-            if (ImGui.Button("Save"))
-            {
-                Plugin.LazyLoot.config.Save();
-                Service.Service.PluginInterface.UiBuilder.AddNotification("Configuration saved", "Lazy Loot", NotificationType.Success);
-            }
-            ImGui.SameLine();
-            if (ImGui.Button("Save and Close"))
-            {
-                Plugin.LazyLoot.ConfigUi.IsOpen = false;
-            }
+            ImGui.Separator();
         }
 
         private static void DrawUserRestriction()
         {
-            ImGui.Text("User Restriction");
             ImGui.Separator();
             ImGui.Checkbox("Ignore item Level below:", ref Plugin.LazyLoot.config.RestrictionIgnoreItemLevelBelow);
             if (Plugin.LazyLoot.config.RestrictionIgnoreItemLevelBelow)
@@ -121,7 +126,45 @@ namespace LazyLoot.Ui
                     Plugin.LazyLoot.config.RestrictionIgnoreItemLevelBelowValue = 0;
                 }
             }
-            ImGui.Checkbox("Ignore items already unlocked. ( Cards, Music, Faded copy, Minions, Mounts )", ref Plugin.LazyLoot.config.RestrictionIgnoreItemUnlocked);
+
+            ImGui.TextColored( ImGuiColors.DalamudRed, "Ignore items even if they are tradeable, do you want to trade them, don't ignore them.");
+            ImGui.Checkbox("Ignore all items already unlocked. ( Cards, Music, Faded copy, Minions, Mounts, Emotes, Hairstyle )", ref Plugin.LazyLoot.config.RestrictionIgnoreItemUnlocked);
+            ImGui.Checkbox("Ignore unlocked mounts.", ref Plugin.LazyLoot.config.RestrictionIgnoreMounts);
+            ImGui.Checkbox("Ignore unlocked minions.", ref Plugin.LazyLoot.config.RestrictionIgnoreMinions);
+            ImGui.Checkbox("Ignore unlocked bardings.", ref Plugin.LazyLoot.config.RestrictionIgnoreBardings);
+            ImGui.Checkbox("Ignore unlocked triple triad cards.", ref Plugin.LazyLoot.config.RestrictionIgnoreTripleTriadCards);
+            ImGui.Checkbox("Ignore unlocked emotes and hairstyle.", ref Plugin.LazyLoot.config.RestrictionIgnoreEmoteHairstyle);
+            ImGui.Checkbox("Ignore unlocked orchestrion rolls.", ref Plugin.LazyLoot.config.RestrictionIgnoreOrchestrionRolls);
+            ImGui.Checkbox("Ignore unlocked Faded Copy.", ref Plugin.LazyLoot.config.RestrictionIgnoreFadedCopy);
+
+            if (Plugin.LazyLoot.config.RestrictionIgnoreMounts
+                && Plugin.LazyLoot.config.RestrictionIgnoreMinions
+                && Plugin.LazyLoot.config.RestrictionIgnoreBardings
+                && Plugin.LazyLoot.config.RestrictionIgnoreTripleTriadCards
+                && Plugin.LazyLoot.config.RestrictionIgnoreEmoteHairstyle
+                && Plugin.LazyLoot.config.RestrictionIgnoreOrchestrionRolls
+                && Plugin.LazyLoot.config.RestrictionIgnoreFadedCopy)
+            {
+                Plugin.LazyLoot.config.RestrictionIgnoreItemUnlocked = true;
+                Plugin.LazyLoot.config.RestrictionIgnoreMounts = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreMinions = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreBardings = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreTripleTriadCards = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreEmoteHairstyle = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreOrchestrionRolls = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreFadedCopy = false;
+            }
+
+            if (Plugin.LazyLoot.config.RestrictionIgnoreItemUnlocked)
+            {
+                Plugin.LazyLoot.config.RestrictionIgnoreMounts = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreMinions = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreBardings = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreTripleTriadCards = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreEmoteHairstyle = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreOrchestrionRolls = false;
+                Plugin.LazyLoot.config.RestrictionIgnoreFadedCopy = false;
+            }
 
             ImGui.Separator();
         }
@@ -130,9 +173,10 @@ namespace LazyLoot.Ui
         {
             ImGui.Checkbox("Display roll information in chat.", ref Plugin.LazyLoot.config.EnableChatLogMessage);
             ImGui.Checkbox("Display roll information as toast:", ref Plugin.LazyLoot.config.EnableToastMessage);
-            ImGui.SameLine();
             if (Plugin.LazyLoot.config.EnableToastMessage)
             {
+                ImGui.SameLine();
+
                 if (Plugin.LazyLoot.config.EnableNormalToast)
                 {
                     toastPreview = "Normal";
@@ -229,14 +273,6 @@ namespace LazyLoot.Ui
 
                 ImGui.EndCombo();
             }
-
-            ImGui.Spacing();
-            ImGui.Text("Delay before Fulf will roll on items. Just to be sure that all chest are open.");
-            ImGui.Text("Doesn't matter most of the time, it's more for stuff like some Normal Raids and Alli Raids.");
-            ImGui.Text("Be careful if you set it too low, Fulf wont roll on all items.");
-            ImGui.DragInt("seconds", ref Plugin.LazyLoot.config.FulfDelay);
-
-            ImGui.Separator();
         }
     }
 }