+++ LazyLoot/Util/RollState.cs
@@ -4,7 +4,7 @@
     {
         UpToNeed = 0,
         UpToGreed = 1,
-        UpToPass = 2,
+        UpToPass = 20,
         Rolled = 17, // 0x00000011
         NoLoot = 26, // 0x0000001A
     }